# docs/swagger.yaml
definitions:
  handlers.APIResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
  models.IDOnly:
    properties:
      id:
        example: 1
        type: integer
    type: object
...
  models.Spread:
    properties:
      description:
        type: string
      id:
        type: integer
      major_arcana:
        type: boolean
      minor_arcana:
        type: boolean
      name:
        type: string
      num_cards:
        type: integer
      upside_down:
        type: boolean
    type: object
  models.SpreadInput:
    properties:
      description:
        type: string
      major_arcana:
        type: boolean
      minor_arcana:
        type: boolean
      name:
        type: string
      num_cards:
        type: integer
      upside_down:
        type: boolean
    type: object
  models.Suit:
    properties:
      description:
        type: string
      genitive:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  models.SuitInput:
    properties:
      description:
        type: string
      genitive:
        type: string
      name:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /spreads:
    get:
      description: Retrieves a list of all available interpretation spreads
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Spread'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.APIResponse'
      summary: Get all spreads
      tags:
      - spreads
    post:
      consumes:
      - application/json
      description: Adds a new spread of interpretations
      parameters:
      - description: Spread data
        in: body
        name: spread
        required: true
        schema:
          $ref: '#/definitions/models.SpreadInput'
      produces:
      - application/json
      responses:
        "201":
          description: Returns the ID of the created spread
          schema:
            $ref: '#/definitions/models.IDOnly'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.APIResponse'
      summary: Create a new spread
      tags:
      - spreads
  /spreads/{id}:
    delete:
      description: Deletes a spread by ID
      parameters:
      - description: Spread ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.APIResponse'
      summary: Delete a spread
      tags:
      - spreads
    get:
      description: Retrieves a spread by its ID
      parameters:
      - description: Spread ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Spread'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.APIResponse'
      summary: Get spread by ID
      tags:
      - spreads
    put:
      consumes:
      - application/json
      description: Updates an existing spread
      parameters:
      - description: Spread ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated spread
        in: body
        name: spread
        required: true
        schema:
          $ref: '#/definitions/models.SpreadInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Spread'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.APIResponse'
      summary: Update a spread
      tags:
      - spreads
  /suits:
    get:
      description: Retrieves a list of all available interpretation suits
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Suit'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.APIResponse'
      summary: Get all suits
      tags:
      - suits
    post:
      consumes:
      - application/json
      description: Adds a new suit of interpretations
      parameters:
      - description: Suit data
        in: body
        name: suit
        required: true
        schema:
          $ref: '#/definitions/models.SuitInput'
      produces:
      - application/json
      responses:
        "201":
          description: Returns the ID of the created suit
          schema:
            $ref: '#/definitions/models.IDOnly'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.APIResponse'
      summary: Create a new suit
      tags:
      - suits
  /suits/{id}:
    delete:
      description: Deletes a suit by ID
      parameters:
      - description: Suit ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.APIResponse'
      summary: Delete a suit
      tags:
      - suits
    get:
      description: Retrieves a suit by its ID
      parameters:
      - description: Suit ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Suit'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.APIResponse'
      summary: Get suit by ID
      tags:
      - suits
    put:
      consumes:
      - application/json
      description: Updates an existing suit
      parameters:
      - description: Suit ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated suit
        in: body
        name: suit
        required: true
        schema:
          $ref: '#/definitions/models.SuitInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Suit'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.APIResponse'
      summary: Update a suit
      tags:
      - suits
swagger: "2.0"